only delete BA

+++ graphql_api/views.py
@@ -26,6 +26,10 @@ class RequestFinalizer:
     A context manager class used as a teardown step after the GraphQL request is fully handled.
     """
 
+    TO_BE_DELETED_PREFIXES = [
+        "bundle_analysis_",
+    ]
+
     def __init__(self):
         pass
 
@@ -36,17 +40,20 @@ class RequestFinalizer:
         """
         folder = "/tmp"
         for filename in os.listdir(folder):
-            file_path = os.path.join(folder, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                log.info(
-                    "Failed to delete temp file",
-                    extra={"file_path": file_path, "exc": e},
-                )
+            if any(
+                filter(filename.startswith, RequestFinalizer.TO_BE_DELETED_PREFIXES)
+            ):
+                file_path = os.path.join(folder, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as e:
+                    log.info(
+                        "Failed to delete temp file",
+                        extra={"file_path": file_path, "exc": e},
+                    )
 
     def __enter__(self):
         pass
